sixth Lesson commit test./

+++ koans/src/beginner/AboutPrimitives.java
@@ -89,8 +89,8 @@ public class AboutPrimitives {
 
     @Koan
     public void bytesHaveASmallerRangeThanShorts() {
-        assertEquals(Byte.MIN_VALUE, -128);
-        assertEquals(Byte.MAX_VALUE, 127);
+        assertEquals(Byte.MIN_VALUE, (byte)-128);
+        assertEquals(Byte.MAX_VALUE, (byte)127);
 
         // Why would you use short or byte considering that you need to do explicit casts?
     }
